removed customer moved in event
