+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.EntryPoints.Outbox/EventHandlers/EventHandlerBase.cs
@@ -1,0 +1,37 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Threading.Tasks;
+using Energinet.DataHub.MarketRoles.Application.Common.Transport;
+using Energinet.DataHub.MarketRoles.Infrastructure.Integration;
+using Energinet.DataHub.MarketRoles.Infrastructure.Transport.Protobuf;
+using Google.Protobuf;
+
+namespace Energinet.DataHub.MarketRoles.EntryPoints.Outbox.EventHandlers
+{
+    public abstract class EventHandlerBase
+    {
+        protected static async Task DispatchMessageAsync<TTopic>(ProtobufOutboundMapper mapper, ITopicSender<TTopic> topicSender, IOutboundMessage request)
+            where TTopic : Topic
+        {
+            if (mapper == null) throw new ArgumentNullException(nameof(mapper));
+            if (topicSender == null) throw new ArgumentNullException(nameof(topicSender));
+
+            var message = mapper.Convert(request);
+            var bytes = message.ToByteArray();
+            await topicSender.SendMessageAsync(bytes).ConfigureAwait(false);
+        }
+    }
+}
