+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.EntryPoints.Outbox/Program.cs
@@ -21,12 +21,11 @@ using Energinet.DataHub.MarketRoles.Domain.SeedWork;
 using Energinet.DataHub.MarketRoles.EntryPoints.Common;
 using Energinet.DataHub.MarketRoles.EntryPoints.Common.MediatR;
 using Energinet.DataHub.MarketRoles.EntryPoints.Common.SimpleInjector;
-using Energinet.DataHub.MarketRoles.EntryPoints.Outbox.Handlers;
+using Energinet.DataHub.MarketRoles.EntryPoints.Outbox.EventHandlers;
 using Energinet.DataHub.MarketRoles.Infrastructure;
-using Energinet.DataHub.MarketRoles.Infrastructure.Correlation;
 using Energinet.DataHub.MarketRoles.Infrastructure.DataAccess;
 using Energinet.DataHub.MarketRoles.Infrastructure.Integration;
-using Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.MoveIn;
+using Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.ChangeOfSupplier;
 using Energinet.DataHub.MarketRoles.Infrastructure.Integration.Services;
 using Energinet.DataHub.MarketRoles.Infrastructure.Outbox;
 using Energinet.DataHub.MarketRoles.Infrastructure.Serialization;
@@ -87,11 +86,6 @@ namespace Energinet.DataHub.MarketRoles.EntryPoints.Outbox
                 Lifestyle.Singleton);
 
             container.Register(
-                () => new ConsumerRegisteredTopic(Environment.GetEnvironmentVariable("CONSUMER_REGISTERED_TOPIC") ?? throw new InvalidOperationException(
-                    "No Consumer Registered Topic found")),
-                Lifestyle.Singleton);
-
-            container.Register(
                 () => new EnergySupplierChangedTopic(Environment.GetEnvironmentVariable("ENERGY_SUPPLIER_CHANGED_TOPIC") ?? throw new InvalidOperationException(
                     "No EnergySupplierChanged Topic found")),
                 Lifestyle.Singleton);
@@ -102,14 +96,15 @@ namespace Energinet.DataHub.MarketRoles.EntryPoints.Outbox
                 new[]
                 {
                     typeof(ConsumerMoveInAccepted).Assembly,
-                    typeof(ConsumerMovedInEvent).Assembly,
+                    typeof(EnergySupplierChangedHandler).Assembly,
                 },
                 Array.Empty<Type>());
 
-            container.AddProtoBuffContracts(
+            container.AddProtobufMessageSerializer();
+            container.AddProtobufOutBoundMappers(
                 new[]
                 {
-                    typeof(ConsumerRegisteredIntegrationEvent).Assembly,
+                    typeof(EnergySupplierChangedIntegrationEvent).Assembly,
                 });
         }
     }
