+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Energinet.DataHub.MarketRoles.Infrastructure.csproj
@@ -53,6 +53,8 @@ limitations under the License.
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Integration\Contracts\MoveIn" />
+    <Folder Include="Integration\IntegrationEventDispatching\MoveIn\Mappers" />
     <Folder Include="Integration\IntegrationEventDispatching\MoveIn\Topics" />
     <Folder Include="Processing\Mappers" />
   </ItemGroup>
