+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Integration/IntegrationEventDispatching/ChangeOfSupplier/EnergySupplierChangedTopic.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.MoveIn
+namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.ChangeOfSupplier
 {
-    public record ConsumerRegisteredTopic(string Name) : Topic;
+    public record EnergySupplierChangedTopic(string Name) : Topic;
 }
