--- source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/Integration/IntegrationEventDispatching/MoveIn/PublishWhenConsumerHasMovedIn.cs
@@ -1,49 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Energinet.DataHub.MarketRoles.Domain.MeteringPoints.Events;
-using Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.MoveIn.Messages;
-using Energinet.DataHub.MarketRoles.Infrastructure.Outbox;
-using MediatR;
-
-namespace Energinet.DataHub.MarketRoles.Infrastructure.Integration.IntegrationEventDispatching.MoveIn
-{
-    public class PublishWhenConsumerHasMovedIn : INotificationHandler<ConsumerMovedIn>
-    {
-        private readonly IOutbox _outbox;
-        private readonly IOutboxMessageFactory _outboxMessageFactory;
-
-        public PublishWhenConsumerHasMovedIn(IOutbox outbox, IOutboxMessageFactory outboxMessageFactory)
-        {
-            _outbox = outbox ?? throw new ArgumentNullException(nameof(outbox));
-            _outboxMessageFactory = outboxMessageFactory ?? throw new ArgumentNullException(nameof(outboxMessageFactory));
-        }
-
-        public Task Handle(ConsumerMovedIn notification, CancellationToken cancellationToken)
-        {
-            if (notification == null) throw new ArgumentNullException(nameof(notification));
-
-            var integrationEvent = new ConsumerMovedInIntegrationEvent(
-                notification.GsrnNumber,
-                notification.MoveInDate);
-
-            var message = _outboxMessageFactory.CreateFrom(integrationEvent, OutboxMessageCategory.IntegrationEvent);
-            _outbox.Add(message);
-            return Task.CompletedTask;
-        }
-    }
-}
